TEP-IFPI-2017_8-master/restaurapp-api-master/requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-enumfields==1.0.0
django-oauth-toolkit==1.2.0
django-rest-passwordreset==0.9.7
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-enum-field==0.9.2
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==2.1.0
openapi-codec==1.3.2
psycopg2-binary==2.7.6.1
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.7
raven==6.9.0
requests==2.20.1
freezegun==0.3.11
django-rest-framework-social-oauth2
simplejson==3.16.0
sqlparse==0.2.4
Unipath==1.1
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
